🚧wip (utils): move constants
