+++ src/pages/api/login.ts
@@ -9,7 +9,7 @@ import { connect } from "db";
 import { comparePassword } from "db/utils";
 import { NotFoundError, routeWrapper } from "api";
 import { ParentModel, SettingsModel, StaffModel, StudentModel } from "db/models";
-import { JWT_ALG, JWT_COOKIE_KEY, USER_ID_COOKIE, USER_LEVEL_COOKIE } from "utils";
+import { JWT_ALG, JWT_COOKIE_KEY, USER_ID_COOKIE, USER_LEVEL_COOKIE } from "utils/constants";
 
 import type { NextApiRequest, NextApiResponse } from "next";
