+++ src/pages/api/students/[id].ts
@@ -2,7 +2,7 @@ import { ReasonPhrases, StatusCodes } from "http-status-codes";
 
 import { connect } from "db";
 import { routeWrapper } from "api";
-import { USER_ID_COOKIE } from "utils";
+import { USER_ID_COOKIE } from "utils/constants";
 import { StudentModel, TeacherStaffModel } from "db/models";
 
 import type { NextApiRequest, NextApiResponse } from "next";
