+++ src/utils/index.ts
@@ -1,20 +1,3 @@
-export const NETWORK_STATE = {
-  online: { emoji: "😄", description: "Found the internet!" },
-  offline: { emoji: "🤔", description: "Hmm... no internet?" },
-};
-
-export const JWT_ALG = "PS256";
-
-export const JWT_COOKIE_KEY = "jwt-k";
-
-export const JWT_COOKIE_TOKEN = "jwt-t";
-
-export const USER_ID_COOKIE = "user-id";
-
-export const USER_LEVEL_COOKIE = "user-level";
-
-export const REDIRECT_QUERY = "to";
-
 /**
  * It takes a country code, converts it to uppercase, splits it into an array of characters, maps each
  * character to its corresponding code point, and then returns a string of the code points
